Split worker from its functions so tests compile
