+++ jest.config.js
@@ -20,10 +20,6 @@ module.exports = {
             diagnostics: {
                 ignoreCodes: [151001],
             },
-            //This can be removed once https://github.com/kulshekhar/ts-jest/issues/1471 is released, probably in ts-jest 25.3.0
-            tsconfig: {
-                outDir: "$$ts-jest$$",
-            },
         },
     },
     testPathIgnorePatterns: ["/node_modules/", "/nightwatch/", "/protobuf/"],
